#50: Refactor to db_manage.py.
