+++ src/raindrop_todoist_syncer/rd_process.py
@@ -2,7 +2,7 @@ from typing import Any
 
 from loguru import logger
 
-from raindrop_todoist_syncer.raindrop import DatabaseManager
+from raindrop_todoist_syncer.db_manage import DatabaseManager
 from raindrop_todoist_syncer.rd_class import Raindrop
